requirements.txt
appdirs==1.4.4
attrs==19.3.0
black==19.10b0
boto==2.49.0
botocore==1.17.23
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
pathspec==0.8.0
praw==7.1.0
prawcore==1.4.0
python-dateutil==2.8.1
python-dotenv==0.14.0
regex==2020.6.8
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
typed-ast==1.4.1
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
